Update unit tests, readme. Add github action

+++ tests/MakeRepositoryCommandTest.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace SmashedEgg\LaravelModelRepository\Tests;
+
 use Orchestra\Testbench\TestCase;
 
 class MakeRepositoryCommandTest extends TestCase
@@ -8,12 +10,12 @@ class MakeRepositoryCommandTest extends TestCase
     {
         parent::tearDown();
 
-        /*if (file_exists(app_path('Repositories/UserRepository.php'))) {
+        if (file_exists(app_path('Repositories/UserRepository.php'))) {
             unlink(app_path('Repositories/UserRepository.php'));
         }
         if (file_exists(app_path('Repositories/AccountRepository.php'))) {
             unlink(app_path('Repositories/AccountRepository.php'));
-        }*/
+        }
     }
 
     public function testCommand()
